Sync DSP, tests, planning corpus, and release tooling
- source: refine InputStage, RateConverterPair, ParameterLayout,
  FixedRateAdapter, SrcLatencyTable; simplify editor/faceplate paint
- tests: expand V1 contract coverage (SrcUnderfill, InputStage, MonoBus,
  PluginBasics, PluginEditor) and tune existing assertions
- planning: add v1.0/v2.0 milestone docs, ADRs, research corpus,
  retrospective, and build microsteps
- scripts: harden macOS package/sign/notarize and verify-v1 flows
- presets: drop stray trailing line from factory XML
- docs: refresh release checklist and RC0 evidence reports
- ci: adjust build_and_test matrix
- gitignore: exclude repomix dumps and vendored repo-samples

+++ tests/HfDiagnosticsHelpers.h
@@ -278,7 +278,7 @@ inline HfMetrics measureTail (const std::string& config,
         }
     }
 
-    m.hfBandDecaySec = static_cast<float> (decaySamples / kSampleRate);
+    m.hfBandDecaySec = static_cast<float> (static_cast<double> (decaySamples) / kSampleRate);
     return m;
 }
